+++ prettier.config.mjs
@@ -1,0 +1,8 @@
+/** @type {import("prettier").Config} */
+const config = {
+  plugins: ["prettier-plugin-tailwindcss"],
+  printWidth: 100,
+  proseWrap: "preserve",
+};
+
+export default config;